Update mpv to 0.34.1 ..FFmpeg to 4.4.2 ..Remove youtube-dl from the repo (use the official yt-dlp binary instead)
