+++ .gitignore
@@ -110,3 +110,6 @@ browser/Firefox_Open_In_IINA/*
 
 *.xcarchive
 /.vscode
+
+deps/executable/youtube-dl
+
