+++ deps/include/libavutil/adler32.h
@@ -27,8 +27,10 @@
 #ifndef AVUTIL_ADLER32_H
 #define AVUTIL_ADLER32_H
 
+#include <stddef.h>
 #include <stdint.h>
 #include "attributes.h"
+#include "version.h"
 
 /**
  * @defgroup lavu_adler32 Adler-32
@@ -37,6 +39,12 @@
  *
  * @{
  */
+
+#if FF_API_CRYPTO_SIZE_T
+typedef unsigned long AVAdler;
+#else
+typedef uint32_t AVAdler;
+#endif
 
 /**
  * Calculate the Adler32 checksum of a buffer.
@@ -50,8 +58,12 @@
  * @param len   size of input buffer
  * @return      updated checksum
  */
-unsigned long av_adler32_update(unsigned long adler, const uint8_t *buf,
-                                unsigned int len) av_pure;
+AVAdler av_adler32_update(AVAdler adler, const uint8_t *buf,
+#if FF_API_CRYPTO_SIZE_T
+                          unsigned int len) av_pure;
+#else
+                          size_t len) av_pure;
+#endif
 
 /**
  * @}
