+++ deps/include/libavutil/ffversion.h
@@ -1,5 +1,5 @@
 /* Automatically generated by version.sh, do not manually edit! */
 #ifndef AVUTIL_FFVERSION_H
 #define AVUTIL_FFVERSION_H
-#define FFMPEG_VERSION "4.2.1"
+#define FFMPEG_VERSION "4.4.2"
 #endif /* AVUTIL_FFVERSION_H */
