+++ deps/include/libavutil/hash.h
@@ -27,6 +27,7 @@
 #ifndef AVUTIL_HASH_H
 #define AVUTIL_HASH_H
 
+#include <stddef.h>
 #include <stdint.h>
 
 #include "version.h"
