+++ deps/include/libavutil/murmur3.h
@@ -27,6 +27,7 @@
 #ifndef AVUTIL_MURMUR3_H
 #define AVUTIL_MURMUR3_H
 
+#include <stddef.h>
 #include <stdint.h>
 
 #include "version.h"
